feat: add Job for sending bills due tomorrow notification for each user
instead of having the logic of veryfying each user's data and based on that send notifcations in a command, now a job handles the logic asynchronously: for each user, a different job is created, placed into a queue to be executed in the background by a queue worker, potentially improving the performance of the application.
